Postdb_tortoise/pyproject.toml
[tool.poetry]
name = "project_with_tortoise"
version = "0.1.0"
description = ""
authors = ["emyrkrkmz <emiryusufkorkmaz2011@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.87.0"
uvicorn = "^0.19.0"
tortoise-orm = "^0.19.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
coverage = "^6.5.0"
pytest-asyncio = "^0.20.2"
httpx = "^0.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
